requirements.txt
streamlit
pyrebase
Pillow
